rework algorithm
rename variables
simplify conditions

Prefilter geohash using polygon.intersects(geohash_polygon).
this check is relatively cheap.
if fully_contained_only is false and they intersect, we're done.
Otherwise, check if the perimiter of the polygon intersects with the
perimeter of the geohash polygon.
If they do, the polygon does NOT fully contain the geohash.

This assumes that the polygon doesn't have "holes" in it.
Support of such polygons can be added.